Server/requirements.txt
pycryptodome
paho-mqtt==1.6.1
